feat: move Access Requests onto the Shadow MCP page
The queue and request detail live where the data they are keyed to
lives: the Shadow MCP page gains Inventory and Access Requests tabs
(inventory keeps its org:admin gate; requests gates on
mcp_approval:read), the request detail moves to
shadow-mcp/requests/:requestId, and the MCP page returns to a plain
servers listing. Breadcrumbs for the requests segment point at the tab.
